refactor: use ModelPath more generally
