+++ scripts/utils.py
@@ -1,8 +1,30 @@
 from pathlib import Path
 
+from pydantic import BaseModel
+
 from scripts.config import classifier_dir
+from src.identifiers import Identifier, WikibaseID
 
 
-def get_local_classifier_path(target_path: str, version: str) -> Path:
+class ModelPath(BaseModel):
+    """Represents the expected path to a model artifact locally or in S3."""
+
+    wikibase_id: WikibaseID
+    classifier_id: Identifier
+
+    def __str__(self) -> str:
+        """
+        Return the path to the model artifact.
+
+        e.g. 'Q123/v4prnc54'
+        """
+        return f"{self.wikibase_id}/{self.classifier_id}"
+
+    def __fspath__(self) -> str:
+        """Return the filesystem path representation for use with pathlib."""
+        return str(self)
+
+
+def get_local_classifier_path(target_path: ModelPath, version: str) -> Path:
     """Returns a path for a classifier file."""
     return classifier_dir / target_path / version / "model.pickle"
